+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 All notable changes to the "fishbone" extension will be documented in this file.
 
-## [0.1.1]
+## [0.2.0]
+- *add nested fishbone* from UI.
+- change title for nested fishbones via UI
 
+## [0.1.1]
 - Fix: 'delete rootcause' deleting wrong element (same index but different category)
 
 ## [0.1.0]
-
 - Initial released version